Launcher3: Move restart check to home page
It's just a flag check. And avoids having to implement a listener for the home key (and still miss home gesture).
Also, restarting the launcher when navigating to previous settings page doesn't seem right.
This way, the launcher is restarted once, if needed, when user goes back to home page, which is where the changes are visible, anyway.

Change-Id: I2e99bb95cbcdb2fdb652e20bb1e9e55b53cab2eb
Signed-off-by: Anush02198 <anush.4376@gmail.com>